requirements/dev.txt
-r ./base.txt
django-extensions==3.1.5
pydot==1.4.2
pyparsing==3.0.6
graphviz==0.18.2